Keep language input on autocompletion
